requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
py-healthcheck==1.10.1
pylint==2.16.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
